SelectionManager update
Updated the selection manager script.
Now it does a radius check around the player and every collider with the tag "Selectable" get highlighted to a redish colour.
When directly looking at an item the color is yellow.
When not looking at the item but are still in radius of it the color returns to the red.

-Added new variables.
-Added hashset.
-Optimized some lines, changed to inline.
-Added drawingGizmo to see the range of the radius in editor.