Image Puzzle
Modified the scripts of the Image puzzle mini game.
Fixed a bug where on tap the image piece would rotate multiple times instead of just once.
Made it so that when all the pieces are in the correct orientation, they become noninteractive to allow the player to view the completed image before the win window pops up.